requirements.txt
SQLAlchemy==2.0.27
python-dotenv==1.0.1
schedule==1.2.1
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiosmtplib==3.0.1
jinja2==3.1.3
gunicorn==21.2.0 